move sanitizeFieldNames into its own util file
